+++ bootstrap/providers.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Core\Audit\Providers\AuditServiceProvider;
 use App\Core\Mail\Providers\MailServiceProvider;
 use App\Core\Settings\Providers\SettingsServiceProvider;
 use App\Providers\AppServiceProvider;
@@ -7,6 +8,7 @@ use App\Providers\Filament\AdminPanelProvider;
 use App\Providers\HorizonServiceProvider;
 
 return [
+    AuditServiceProvider::class,
     MailServiceProvider::class,
     SettingsServiceProvider::class,
     AppServiceProvider::class,